Python Interface update
Minor changes in the C++ library (elimination of warnings for new compilers)
Update of build scripts for successful build on Windows 10 + Visual Studio 2017 + Anaconda
New Python example for jointmv models

+++ libclandmark/CDeformationCost.h
@@ -151,7 +151,8 @@ protected:
 	// Loss function
 	CLoss *loss;
 
-	int kDimension;
+	//int kDimension;
+	size_t kDimension;
 
 	// internal representation of features goes to specialized classes (e.g. CDisplacementDeformationCost, etc.)
 	// convention on order for edge (parent, child)
